backend/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.133.1",
    "uvicorn>=0.41.0",
    "python-dotenv>=1.2.1"
]
